// packages/json-joy/src/json-patch/codec/compact/types.ts
export type OPCODE_TYPE = OPCODE.type | 'type';
// ...
export type CompactTypeOp = [opcode: OPCODE_TYPE, path: string | Path, value: JsonPatchTypes];